+++ sql/schema/03_sets.sql
@@ -1,7 +1,7 @@
 -- +goose Up
 CREATE TABLE sets (
     id BIGSERIAL PRIMARY KEY,
-    set_order SMALLINT NOT NULL,
+    set_order INTEGER NOT NULL,
     rest_time INTEGER,
     session_id UUID NOT NULL,
     CONSTRAINT fk_session_id FOREIGN KEY (session_id)
